requirements.txt
# FastAPI and API dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM inference
torch==2.1.2
transformers==4.46.0
tokenizers==0.20.3
accelerate==0.26.1
bitsandbytes==0.42.0

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-logging==0.43b0

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.7

# Load testing
locust==2.20.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
